[Design] #28 - detail header 구현
